Employee Department Re Check Submit

+++ app/Http/Controllers/Admin/Configuration/Employee/EmployeeController.php
@@ -7,12 +7,17 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Yajra\Datatables\Datatables;
 use App\models\employee\Department;
+use Illuminate\Validation\Rule;
 
 class EmployeeController extends Controller
 {
 
     public function Department_index()
     {
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
+        
         return view('admin.employee.department.index');
     }
 
@@ -24,22 +29,33 @@ class EmployeeController extends Controller
             $department = Department::where('name', '!=', config('system.default_role.admin'))->get();
             return Datatables::of($department)
                 ->addIndexColumn()
+                ->editColumn('description', function($model) {
+                    return str_limit($model->description, 60);
+                })
                 ->addColumn('action', function ($model) {
                     return view('admin.employee.department.action', compact('model'));
-                })->rawColumns(['action'])->make(true);
+                })->rawColumns(['action', 'description'])->make(true);
         }
     }
 
     public function Department_create()
     {
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('admin.employee.department.create');
     }
 
 
     public function Department_store(Request $request)
     {
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $validator = $request->validate([
-            'name' => 'required|min:3|unique:departments,name,NULL,id,deleted_at,NULL',
+            'name' => 'required|min:3|max:50|unique:departments,name,NULL,id,deleted_at,NULL',
             'description' => '',
         ]);
 
@@ -60,6 +76,10 @@ class EmployeeController extends Controller
 
     public function Department_edit($id)
     {
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         // find the data
         $model = Department::where('id', $id)->firstOrFail();
 
@@ -69,12 +89,16 @@ class EmployeeController extends Controller
 
     public function Department_update(Request $request, $id)
     {
-        $validator = $request->validate([
-            'name' => 'required|max:255|unique:departments,name,'.$id,
-            'description' => '',
-        ]);
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
 
         $model = Department::findOrFail($id);
+
+        $request->validate([
+            'name' => 'required|min:3|max:50|unique:departments,name,NULL,id,deleted_at,NULL'.$id,
+        ]);
+
         $model->name = $request->name;
         $model->description = $request->description;
         $model->save();
@@ -86,6 +110,10 @@ class EmployeeController extends Controller
 
     public function Department_delete(Request $request, $id)
     {
+        if (!auth()->user()->can('workorder.update')) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $type = Department::findOrFail($id);
         $name = $type->name;
         $type->delete();
